# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lsproxy",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.5"
  },
  "servers": [
    {
      "url": "http://localhost:4444/v1",
      "description": "API server v1"
    }
  ],
  "paths": {
    "/symbol/definitions-in-file": {
      "get": {
        "tags": [
          "symbol"
        ],
        "summary": "Get symbols in a specific file (uses ast-grep)",
        "description": "Returns a list of symbols (functions, classes, variables, etc.) defined in the specified file.\n\nOnly the variabels defined at the file level are included.\n\nThe returned positions point to the start of the symbol's identifier.\n\ne.g. for `User` on line 0 of `src/main.py`:\n```\n0: class User:\n_________^\n1:     def __init__(self, name, age):\n2:         self.name = name\n3:         self.age = age\n```",
        "operationId": "definitions_in_file",
        "parameters": [
          {
            "name": "file_path",
            "in": "query",
            "description": "The path to the file to get the symbols for, relative to the root of the workspace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Symbols retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Symbol"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/symbol/find-identifier": {
      "post": {
        "tags": [
          "symbol"
        ],
        "summary": "Finds occurrences of an identifier by name in a file",
        "description": "Given a file path and identifier name, returns:\n- Without position: All matching identifiers in the file\n- With position: The exact identifier with that name at that position, or 3 closest identifiers with that name\n\nExample finding all occurrences of \"user_name\":\n```\nlet user_name = \"John\";  // First occurrence\nprintln!(\"{}\", user_name); // Second occurrence\n```\n\nWhen a position is provided, it searches for an exact match at that location.\nIf no exact match exists, returns the 3 identifiers closest to the position\nbased on line and character distance, prioritizing lines.",
        "operationId": "find_identifier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindIdentifierRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Identifier retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentifierResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FilePosition": {
        "type": "object",
        "description": "Specific position within a file.",
        "required": [
          "path",
          "position"
        ],
        "properties": {
          "path": {
            "type": "string",
            "example": "src/main.py"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          }
        }
      },
      "FileRange": {
        "type": "object",
        "required": [
          "path",
          "start",
          "end"
        ],
        "properties": {
          "end": {
            "$ref": "#/components/schemas/Position",
            "description": "The end position of the range."
          },
          "path": {
            "type": "string",
            "description": "The path to the file.",
            "example": "src/main.py"
          },
          "start": {
            "$ref": "#/components/schemas/Position",
            "description": "The start position of the range."
          }
        }
      },
...
      "FindIdentifierRequest": {
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the identifier to search for.",
            "example": "User"
          },
          "path": {
            "type": "string",
            "description": "The path to the file to search for identifiers.",
            "example": "src/main.py"
          },
          "position": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Position",
                "description": "The position hint to search for identifiers. If not provided."
              }
            ]
          }
        }
      },
...
      "Identifier": {
        "type": "object",
        "required": [
          "name",
          "range"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "range": {
            "$ref": "#/components/schemas/FileRange"
          }
        }
      },
      "IdentifierResponse": {
        "type": "object",
        "required": [
          "identifiers"
        ],
        "properties": {
          "identifiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Identifier"
            }
          }
        }
      },
      "Position": {
        "type": "object",
        "required": [
          "line",
          "character"
        ],
        "properties": {
          "character": {
            "type": "integer",
            "format": "int32",
            "description": "0-indexed character index.",
            "example": 5,
            "minimum": 0
          },
          "line": {
            "type": "integer",
            "format": "int32",
            "description": "0-indexed line number.",
            "example": 10,
            "minimum": 0
          }
        }
      },
...
      "Symbol": {
        "type": "object",
        "required": [
          "name",
          "kind",
          "identifier_position",
          "range"
        ],
        "properties": {
          "identifier_position": {
            "$ref": "#/components/schemas/FilePosition",
            "description": "The start position of the symbol's identifier."
          },
          "kind": {
            "type": "string",
            "description": "The kind of the symbol (e.g., function, class).",
            "example": "class"
          },
          "name": {
            "type": "string",
            "description": "The name of the symbol.",
            "example": "User"
          },
          "range": {
            "$ref": "#/components/schemas/FileRange",
            "description": "The full range of the symbol."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}